requirements.txt

# Required Python packages for the integrated ForestNAV application
PyQt5
PyQtWebEngine
folium
geopandas
shapely
pandas
numpy
matplotlib
seaborn
fiona
chardet
pytz
pyserial
requests